// src/types/cart.ts
export type Cart = {
  attrsText: string;
  count: number;
  discount: number;
  id: string;
  isCollect: boolean;
  isEffective: boolean;
  name: string;
  nowOriginalPrice: string;
  nowPrice: string;
  picture: string;
  postFee: number;
  price: string;
  selected: boolean;
  skuId: string;
  specs: string[];
  stock: number;
};